change timer remove images

+++ src/job/RemoveImagesLocalJob.ts
@@ -11,33 +11,35 @@ export class RemoveImagesLocalJob extends BaseJob {
   }
 
   async job(): Promise<any> {
-    const t0 = performance.now();
-    const filePath = `${_pathFile}`;
-    const newDate = new Date().getTime();
-    const twoDay = 17280000;
-    try {
-      fs.readdir(filePath, (err: NodeJS.ErrnoException | null, files: string[]) => {
-        if (err) {
-          throw err;
-        }
-        for (let file of files) {
-          if (parseInt(file.split('.')[0]) + twoDay < newDate) {
-            fs.unlink(`${filePath}/${file}`, (err) => {
-              if (err) {
-                logger.error('RemoveImagesLocalJob error', {
-                  error: `can not remove file:::: ${file}`
-                });
-              }
-              console.log(`RemoveImagesLocalJob: hass remove file:::: ${file}`);
-            });
+    if (process.env.STATUS_REMOVE_LOCAL !== 'OFF') {
+      const t0 = performance.now();
+      const filePath = `${_pathFile}`;
+      const newDate = new Date().getTime();
+      const twoDay = 17280000;
+      try {
+        fs.readdir(filePath, (err: NodeJS.ErrnoException | null, files: string[]) => {
+          if (err) {
+            throw err;
           }
-        }
-      });
-    } catch (error) {
-      logger.error('RemoveImagesLocalJob error', { error: error });
-      return error;
+          for (let file of files) {
+            if (parseInt(file.split('.')[0]) + twoDay < newDate) {
+              fs.unlink(`${filePath}/${file}`, (err) => {
+                if (err) {
+                  logger.error('RemoveImagesLocalJob error', {
+                    error: `can not remove file:::: ${file}`
+                  });
+                }
+                console.log(`RemoveImagesLocalJob: hass remove file:::: ${file}`);
+              });
+            }
+          }
+        });
+      } catch (error) {
+        logger.error('RemoveImagesLocalJob error', { error: error });
+        return error;
+      }
+      const t1 = performance.now();
+      logger.info(`RemoveImagesLocalJob: Ending RemoveImagesLocalJob`, { time: t1 - t0 });
     }
-    const t1 = performance.now();
-    logger.info(`RemoveImagesLocalJob: Ending RemoveImagesLocalJob`, { time: t1 - t0 });
   }
 }
